+++ app/src/main/java/woowacourse/movie/model/movie/screening/DefaultShowtimePolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model.movie.screening
 
 import java.time.DayOfWeek
 import java.time.LocalDate
