+++ app/src/main/java/woowacourse/movie/model/movie/screening/Screening.kt
@@ -1,5 +1,6 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model.movie.screening
 
+import woowacourse.movie.model.movie.Movie
 import java.time.LocalDate
 import java.time.LocalTime
