+++ app/src/main/java/woowacourse/movie/model/movie/screening/ShowtimePolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model.movie.screening
 
 import java.time.LocalDate
 import java.time.LocalTime
